services/users/requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
psycopg2==2.9.5
flask-testing==0.8.1
gunicorn==20.1.0
coverage==7.2.1
flake8===6.0.0
flask-debugtoolbar==0.13.1
flask-cors==3.0.10
flask-migrate==4.0.4
flask-bcrypt==1.0.1
PyJWT==2.6.0